+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/cache.php
@@ -16,7 +16,6 @@ use Symfony\Component\Cache\Adapter\AbstractAdapter;
 use Symfony\Component\Cache\Adapter\AdapterInterface;
 use Symfony\Component\Cache\Adapter\ApcuAdapter;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
-use Symfony\Component\Cache\Adapter\DoctrineAdapter;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Cache\Adapter\MemcachedAdapter;
 use Symfony\Component\Cache\Adapter\PdoAdapter;
@@ -93,28 +92,7 @@ return static function (ContainerConfigurator $container) {
             ->call('setLogger', [service('logger')->ignoreOnInvalid()])
             ->tag('cache.pool', ['clearer' => 'cache.default_clearer', 'reset' => 'reset'])
             ->tag('monolog.logger', ['channel' => 'cache'])
-        ;
 
-    if (class_exists(DoctrineAdapter::class)) {
-        $container->services()->set('cache.adapter.doctrine', DoctrineAdapter::class)
-            ->abstract()
-            ->args([
-                abstract_arg('Doctrine provider service'),
-                '', // namespace
-                0, // default lifetime
-            ])
-            ->call('setLogger', [service('logger')->ignoreOnInvalid()])
-            ->tag('cache.pool', [
-                'provider' => 'cache.default_doctrine_provider',
-                'clearer' => 'cache.default_clearer',
-                'reset' => 'reset',
-            ])
-            ->tag('monolog.logger', ['channel' => 'cache'])
-            ->deprecate('symfony/framework-bundle', '5.4', 'The abstract service "%service_id%" is deprecated.')
-        ;
-    }
-
-    $container->services()
         ->set('cache.adapter.filesystem', FilesystemAdapter::class)
             ->abstract()
             ->args([